requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
celery==5.5.2
gunicorn==23.0.0
drf-yasg==1.21.10
django-cors-headers==4.7.0
python-dotenv==1.0.1
redis==6.0.0
